remove common module and move contents to top-level
