+++ qiskit/pulse/commands/instruction.py
@@ -12,8 +12,8 @@ import logging
 from copy import copy
 from typing import Tuple
 
-from qiskit.pulse.common.interfaces import ScheduleComponent
-from qiskit.pulse.common.timeslots import TimeslotCollection
+from qiskit.pulse.interfaces import ScheduleComponent
+from qiskit.pulse.timeslots import TimeslotCollection
 from .pulse_command import PulseCommand
 
 logger = logging.getLogger(__name__)
