--- qiskit/pulse/common/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Copyright 2019, IBM.
-#
-# This source code is licensed under the Apache License, Version 2.0 found in
-# the LICENSE.txt file in the root directory of this source tree.
-
-"""Common classes for pulse."""
