+++ qiskit/pulse/timeslots.py
